requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
auditlog==0.0.5
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.1.6
django-decouple==2.1
django-filter==21.1
django-guardian==2.4.0
django-import-export==2.6.1
django-jet-reboot==1.1.1
django-jsonfield==1.4.1
django-phonenumber-field==5.2.0
django-smart-selects==1.5.9
django-tinymce4-lite==1.8.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
feedparser==6.0.8
google-api-python-client==1.4.1
httpie==2.6.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.2
install==1.3.5
jsmin==3.0.0
kombu==5.2.2
lxml==4.7.1
Markdown==3.3.6
MarkupPy==1.14
oauth2client==4.1.3
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.36
Pillow==8.4.0
pkg_resources==0.0.0
prompt-toolkit==3.0.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.6.0
python-docx==0.8.11
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
